telegramm_bot/core/database/engine.py
import os
from dotenv import load_dotenv
from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine

from .models import Base

load_dotenv(dotenv_path="/home/vvv/Python/scraper_news/Scraper_News/.env")
engine = create_async_engine(os.getenv("DB_SQLITE"), echo=True)

session_maker = async_sessionmaker(bind=engine, class_=AsyncSession, expire_on_commit=False)#class_=AsyncSession - создание ассинхр сессий

async def create_db():
    async with engine.begin() as conn:
        await conn.run_sync(Base.metadata.create_all)

async def drop_db():
    async with engine.begin() as conn:
        await conn.run_sync(Base.metadata.drop_all)