requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp_socks==0.10.1
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-ai-documentintelligence==1.0.2
azure-core==1.34.0
azure-identity==1.23.0
beautifulsoup4==4.13.4
BingImageCreator==0.5.0
blinker==1.9.0
bottle==0.13.3
Brotli==1.1.0
browser-cookie3==0.20.1
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cobble==0.1.4
coloredlogs==15.0.1
cryptography==45.0.3
cssselect2==0.8.0
curl_cffi==0.11.1
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
duckduckgo_search==8.0.2
EbookLib==0.19
et_xmlfile==2.0.0
fake-useragent==2.2.0
fastapi==0.115.12
Flask==3.1.1
flatbuffers==25.2.10
frozenlist==1.6.0
g4f==0.5.3.1
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.10.0
lxml==5.4.0
lz4==4.4.4
magic-filter==1.0.12
magika==0.6.2
mammoth==1.9.0
markdownify==1.1.0
markitdown==0.1.1
MarkupSafe==3.0.2
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
mss==10.0.0
multidict==6.4.4
nest-asyncio==1.6.0
nodriver==0.46.1
numpy==2.2.6
odfpy==1.4.1
olefile==0.47
onnxruntime==1.22.0
openai==1.81.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pdfminer.six==20250506
pillow==11.2.1
platformdirs==4.3.8
plyer==2.1.0
primp==0.15.0
propcache==0.3.1
protobuf==6.31.0
proxy_tools==0.1.0
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==1.0.2
python-socks==2.7.1
pytz==2025.2
pywebview==5.4
regex==2024.11.6
requests==2.32.3
selenium==4.32.0
selenium-stealth==1.0.6
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
tinycss2==1.4.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.3
yarl==1.20.0
youtube-transcript-api==1.0.3
